Protect against installs on less than Cinnamon 5.2

+++ scripts/bin/dd_wake.sh
@@ -1,4 +1,11 @@
 #!/usr/bin/env bash
+
+CINN_VERSION=$(cinnamon --version)
+if awk "BEGIN {exit !($CINN_VERSION < 5.2)}"; then
+    echo "[i] ERROR: Cinnamon version too low, this script was designed for Cinnamon 5.2 and above"
+    exit
+fi
+
 rm -f $HOME/.cache/wallpaper_current.txt
 touch $HOME/.cache/wallpaper_current.txt
